% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_decoding_plots.R
\name{create_decoding_plots}
\alias{create_decoding_plots}
\title{Visualize decoded hidden states}
\usage{
create_decoding_plots(list_results, method = "filtering", print = FALSE)
}
\arguments{
\item{list_results}{An object of class \code{epiquest_decoding} produced by \code{run_out_of_sample_decoding()}.}

\item{method}{A character string specifying the decoding method to visualize.
Must be one of \code{'filtering'} (default), \code{'local'}, or \code{'global'}.}

\item{print}{A logical. If \code{TRUE}, all generated plots are printed to the
active graphics device.}
}
\value{
An object of class \code{epiquest_plot_list}, a named \code{list} of \code{ggplot2} objects containing:
\itemize{
\item \code{time_series}: A ggplot object with points colored
by the most likely state.
\item \code{time_series_shaded}: A ggplot object similar to \code{time_series},
but where point transparency represents state probability
(not produced if \code{method = 'global'}).
}
}
\description{
Generates two plots to visualize state assignments for (out-of-sample) state decoding:
one with solid colored points (fixed state assignment) and another where point
transparency (alpha) is weighted by the decoding probability. The data can be
\emph{out-of-sample}, meaning that they are new data on which the hidden Markov model was
not trained.
}
\details{
Consult \code{\link{run_out_of_sample_decoding}} for information about the different
decoding options for method. The vertical dotted line marks the last index of the training
data for \code{hmm_frozen}. Data points to its right are the new out-of-sample data.

The plots generated by \code{create_hmm_plots()} and \code{create_loop_plots()} all use \code{'local'}
state assignments.
}
\seealso{
\code{\link{run_out_of_sample_decoding}}
}
