% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_sari_be.R
\docType{data}
\name{df_sari_be}
\alias{df_sari_be}
\title{Belgian SARI data}
\format{
\subsection{\code{df_sari_be}}{

A data frame with 206 rows and 2 columns:
\describe{
\item{index}{Week encoded as Monday of that week (Date variable)}
\item{rate}{Belgian SARI incidence per 100,000 inhabitants (numerical variable)}
}
}
}
\source{
\url{https://www.sciensano.be/en/projects/severe-acute-respiratory-infection-surveillance-a-sentinel-network-hospitals}
}
\usage{
df_sari_be
}
\description{
Sciensano, the Belgian Institute for Health, coordinates a surveillance network of hospitals
that record all cases of severe acute respiratory infection (SARI). These data are used to estimate
the weekly incidence of SARI hospitalisations in Belgium.
}
\details{
This data set contains synthetic data that mimics the behavior of the real estimated incidences
for the period 2021-06-21 to 2025-05-26.
}
\keyword{datasets}
