% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ggplot_theme.R
\name{generate_ggplot_theme}
\alias{generate_ggplot_theme}
\title{Standardized ggplot2 theme for EpiQUEST}
\usage{
generate_ggplot_theme()
}
\value{
A \code{ggplot2} theme object.
}
\description{
Provides a consistent visual style for all package visualizations,
ensuring uniform font sizes and legend placement.
}
\keyword{internal}
