% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palette_hmm.R
\name{generate_palette_hmm}
\alias{generate_palette_hmm}
\title{Generate standardized ggplot2 color scales for HMM states}
\usage{
generate_palette_hmm(n_states = 3, fill = FALSE)
}
\arguments{
\item{n_states}{An integer. The number of hidden states (2 to 4).}

\item{fill}{A logical. If \code{TRUE}, returns \code{scale_fill_manual}.
If \code{FALSE} (default), returns \code{scale_color_manual}.}
}
\value{
A \code{ggplot2} manual color/fill scale object.
}
\description{
Generates a manual color or fill scale for \code{ggplot2} using a
pre-defined color palette designed for epidemiological surveillance signals.
}
\details{
The palette uses specific hex codes to represent severity. In general:
\itemize{
\item \strong{Green} : (very) low activity.
\item \strong{Yellow}: medium activity.
\item \strong{Red}: high activity.
}
}
\seealso{
\code{\link{generate_state_labels_hmm}}
}
\keyword{internal}
