% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epiquest_hmm.R
\name{plot.epiquest_hmm}
\alias{plot.epiquest_hmm}
\title{Plot method for HMM results}
\usage{
\method{plot}{epiquest_hmm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{epiquest_hmm} produced by \code{run_hmm()}.}

\item{...}{Additional arguments passed to methods (not currently used).}
}
\value{
A \code{ggplot2} object showing the time series colored by the
most probable hidden state.
}
\description{
A wrapper for \code{\link{create_hmm_plots}} that provides the
"full time series" visualization by default.
}
\keyword{internal}
