% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_epiquest_plot_list.R
\name{print.epiquest_plot_list}
\alias{print.epiquest_plot_list}
\title{Print method for EpiQUEST plot lists}
\usage{
\method{print}{epiquest_plot_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{epiquest_plot_list}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Ensures that when a list of plots is returned to the console,
the plots are drawn automatically without printing the list structure.
}
\keyword{internal}
