% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hmm_rate.R
\name{run_hmm_rate}
\alias{run_hmm_rate}
\title{Fit HMM for continuous rate data}
\usage{
run_hmm_rate(obs_data, n_states = 2, seasonal = FALSE)
}
\arguments{
\item{obs_data}{A \code{data.frame} containing an \code{index} column (Date
or integer). Additional required columns depend on \code{type} and \code{seasonal}:
\itemize{
\item If \code{type = "rate"}: Must contain a numeric \code{rate} column.
This represents the intensity, incidence or activity level of the
surveillance signal (e.g., cases per 100,000 population).
\item If \code{type = "perc"}: Must contain integer \code{num} and
\code{denom} columns. \code{num} is the numerator (e.g., the number
of positive lab tests or symptomatic patients), and \code{denom}
is the denominator or total sample size (e.g., total tests
performed or total clinical consultations).
\item If \code{seasonal = TRUE}: Must contain a \code{season} column.
The subseries in each season is treated as an independent time series.
}}

\item{n_states}{An integer. The number of hidden states (2 to 4).}

\item{seasonal}{A logical. If \code{TRUE}, the model prevents transitions
between different seasons, treating them as independent sequences.}
}
\value{
An object of class \code{epiquest_hmm}.
}
\description{
Internal function to fit a Gaussian HMM to rate time series.
It uses the Baum-Welch algorithm for parameter estimation and provides
smoothed state probabilities.
}
\details{
The model assumes a Gaussian (normal) distribution for the response variable
\code{rate}. If \code{seasonal} is \code{TRUE}, the \code{ntimes} parameter
in \code{depmixS4} is populated with the lengths of each individual season
to prevent transitions between the end of one season and the start of the next.
}
\keyword{internal}
