% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_out_of_sample_decoding.R
\name{run_out_of_sample_decoding}
\alias{run_out_of_sample_decoding}
\title{Out-of-sample state decoding using a frozen HMM}
\usage{
run_out_of_sample_decoding(obs_data_new, hmm_frozen, seasonal = FALSE)
}
\arguments{
\item{obs_data_new}{A \code{data.frame} containing the new observations.
Must contain an \code{index} column and the same response columns
(e.g., \code{rate} or \code{num}/\code{denom}) used to fit the \code{hmm_frozen}.
If \code{seasonal = TRUE}, it must also contain a \code{season} column.}

\item{hmm_frozen}{An object of class \code{epiquest_hmm}. This is the
fitted model object containing the parameters to be applied to the new data.}

\item{seasonal}{A logical. If \code{TRUE}, the function respects season
boundaries in \code{obs_data_new} to prevent transitions between different
seasons.}
}
\value{
An object of class \code{epiquest_decode}. This is a list containing:
\itemize{
\item \code{states}: The input \code{obs_data_new} with added columns for
the predicted state and method used (see Details below).
\item \code{probs}: For methods 'local' and 'filtering', the posterior probabilities
for each of the states.
}
}
\description{
This function applies the parameters of a previously fitted hidden Markov model
(an \code{epiquest_hmm} object) to new data. It allows for
freezing of the HMM parameters (transition probabilities, emission
distributions and initial state distribution) and determining the hidden
states for new data without re-estimating the model.
}
\details{
The function initializes a new \code{depmix} model using the structure of
\code{obs_data_new} and then uses \code{depmixS4::setpars} to hard-code the
parameters from \code{hmm_frozen}.

This is particularly useful for prospective surveillance, where you want to
classify new weekly data using a model trained on historical seasons without
the new data points influencing the previously trained model.

The states can be decoded according to different methods:
\itemize{
\item \code{'global'}: Uses the Viterbi algorithm to find the single
\bold{most likely sequence} of states across the entire time series. This approach
outputs a single state for each data point. There are no associated probabilities.
\item \code{'local'}: Assigns the state with the \bold{highest posterior
smoothing probability} at each specific time point.
\item \code{'filtering'}: Similar to 'local', but performs real-time decoding,
where the state at time \emph{t} is determined only by observations from
time 1 to \emph{t}.
}
Comparing 'local' and 'filtering', each approach has its advantages and disadvantages.
Filtering state assignments do not use information from the following week to determine states,
so they do not use all available information. When computed every week on new data as the data
arrive, the state assignments are stable. Local state assingments can change: they are only stable
once the data for the following week is available.

The plots generated by \code{create_hmm_plots()} and \code{create_loop_plots()} all use \code{'local'}
state assignments.
}
