% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_threshold_computation.R
\name{run_threshold_computation}
\alias{run_threshold_computation}
\title{Compute QUEST thresholds}
\usage{
run_threshold_computation(
  list_results,
  quantiles = c(0.05, 0.7, 0.9, 0.99),
  epidemic_state_indices = NULL
)
}
\arguments{
\item{list_results}{An object of class \code{epiquest_hmm} produced by \code{run_hmm()}.}

\item{quantiles}{A numeric vector. The cumulative probabilities (between 0 and 1)
at which to calculate thresholds. Defaults to \code{c(0.05, 0.70, 0.90, 0.99)}.}

\item{epidemic_state_indices}{An integer vector. The indices of the HMM states
to be jointly considered as the "epidemic" signal. If \code{NULL},
defaults to the state with the highest mean (see section
'Defining the "epidemic" state(s)' below).}
}
\value{
An object of class \code{epiquest_thresholds}, which inherits from
\code{epiquest_hmm}. This is the input \code{list_results} with 3
additional elements:
\itemize{
\item \code{thresholds}: A numeric vector of calculated rate values.
\item \code{quantiles}: The input quantile levels used for calculation.
\item \code{epidemic_state_indices}: The indices of states used to define
the epidemic state.
}
}
\description{
Calculates QUEST thresholds (default: low, medium, high, very high)
by determining specific quantiles of the observed surveillance signal, weighted by the
posterior probability that the system is in a user-defined epidemic state.
}
\details{
This function implements a "soft" thresholding approach. Instead of calculating
quantiles from the raw data alone, it uses the HMM posterior probabilities
to weight the observations.

For each observation, the function:
\enumerate{
\item Determines how likely a specific week is to be in the "epidemic" state(s).
\item Constructs a weighted empirical cumulative distribution function (ECDF)
of the observed \code{rate}.
\item Uses linear interpolation to find the exact rates that correspond to your
chosen \code{quantiles}.
}
}
\section{Defining the "epidemic" state(s)}{

The \code{epidemic_state_indices} argument allows you to define what states
in the HMM output represent "epidemic activity". This flexibility is useful in
several scenarios:
\itemize{
\item \strong{Combining states}: In a 3-state model, you might consider
both state 2 ("Elevated") and state 3 ("Epidemic") to represent
high activity activity. You would set \code{epidemic_state_indices = c(2, 3)}.
\item \strong{Excluding extreme outliers}: If the data includes a
"super-epidemic" year, the highest HMM state might capture
only those rare extremes. To set thresholds that are more sensitive
to normal annual epidemics, you might choose to define the epidemic
signal based only on the state with the second highest mean.
}

If no indices are provided, the function defaults to the state with the
highest mean incidence.
}

\examples{
# Fit a 3-state HMM to (continuous) rate data
fit <- run_hmm(df_sari_be, n_states = 3, type = "rate")

# Check state information
summary(fit)

# Visualize state information
create_hmm_plots(fit)

# Compute QUEST thresholds using the highest state (L3) as the epidemic
# state. By default, epidemic_state_indices is the highest state.
thresh <- run_threshold_computation(fit)

# Check QUEST threshold information
summary(thresh)

# Visualize threshold information
create_threshold_plots(thresh)
}
