% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_epiquest_hmm.R
\name{summary.epiquest_hmm}
\alias{summary.epiquest_hmm}
\title{Summary of hidden Markov model fit}
\usage{
\method{summary}{epiquest_hmm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{epiquest_hmm} produced by \code{run_hmm()}.}

\item{...}{Additional arguments passed to methods (not currently used).}
}
\value{
The function prints a summary to the console and invisibly
returns the \code{object}.
}
\description{
Provides a readable summary of an \code{epiquest_hmm} object,
including model type, state definitions, and transition probabilities.
}
\details{
The last section on state distribution (observations) does not count the number
of observations in each state since the model does not make hard state assignments.
Rather, it provides for each observations a (posterior smoothing) probability that
it is in each of the states. In other words, each observation is assigned a weight
of 1 that is distributed over the different states. The summary provides the total
weight in each state.
}
