% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_epiquest_loop.R
\name{summary.epiquest_loop}
\alias{summary.epiquest_loop}
\title{Summary of HMM stability analysis}
\usage{
\method{summary}{epiquest_loop}(object, ...)
}
\arguments{
\item{object}{An object of class \code{epiquest_loop}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The function prints a summary to the console and invisibly
returns the \code{object}.
}
\description{
Provides a readable summary of the stability of thresholds and HMM
state means over multiple iterative refits (the loop).
}
